reorganise the headers, and add templated getter/setters
